// src/shared/types/milionario-tips.d.ts
export type MilionarioTipsLatestMatchesChamp = {
    Id: number;
    competition: string;
    changeDate: string;
    matches: {
        Id: number;
        TeamHome: string;
        TeamAway: string;
        SumScore: number;
        FinalTimeResult: string;
        HalfTimeResult: string;
        Date: string;
        HomeImg: string | null;
        AwayImg: string | null;
    }[]
}